--- src/main/java/nextstep/auth/authentication/TokenAuthenticationInterceptor2.java
@@ -1,71 +1,0 @@
-package nextstep.auth.authentication;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.context.Authentication;
-import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenResponse;
-import nextstep.member.application.CustomUserDetailsService;
-import nextstep.member.domain.LoginMember;
-import org.springframework.http.MediaType;
-import org.springframework.web.servlet.HandlerInterceptor;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class TokenAuthenticationInterceptor2 implements HandlerInterceptor {
-
-    private final CustomUserDetailsService customUserDetailsService;
-    private final JwtTokenProvider jwtTokenProvider;
-    private final AuthenticationConverter authenticationConverter;
-
-    public TokenAuthenticationInterceptor2(CustomUserDetailsService customUserDetailsService,
-                                           JwtTokenProvider jwtTokenProvider,
-                                           AuthenticationConverter authenticationConverter) {
-        this.customUserDetailsService = customUserDetailsService;
-        this.jwtTokenProvider = jwtTokenProvider;
-        this.authenticationConverter = authenticationConverter;
-    }
-
-    @Override
-    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
-        AuthenticationToken authenticationToken = authenticationConverter.convert(request);
-        Authentication authentication = authenticate(authenticationToken);
-
-        String token = token(authentication);
-        TokenResponse tokenResponse = new TokenResponse(token);
-
-        String responseToClient = new ObjectMapper().writeValueAsString(tokenResponse);
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.getOutputStream().print(responseToClient);
-
-        return false;
-    }
-    public Authentication authenticate(AuthenticationToken authenticationToken) {
-        String principal = authenticationToken.getPrincipal();
-        LoginMember loginMember = customUserDetailsService.loadUserByUsername(principal);
-        checkAuthentication(loginMember, authenticationToken);
-
-        return new Authentication(loginMember);
-    }
-
-    private String token(Authentication authentication) throws JsonProcessingException {
-        ObjectMapper mapper = new ObjectMapper();
-        String payload = mapper.writeValueAsString(authentication.getPrincipal());
-
-        return jwtTokenProvider.createToken(payload);
-    }
-
-    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
-        if (userDetails == null) {
-            throw new AuthenticationException();
-        }
-
-        if (!userDetails.checkPassword(token.getCredentials())) {
-            throw new AuthenticationException();
-        }
-    }
-
-}
